+++ src/plugins.rs
@@ -1,5 +1,5 @@
+pub mod code;
 pub mod doctor;
-pub mod git;
 pub mod net;
 pub mod tv;
 pub mod yt;