+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add linting system
 - Add preflight checks and remove git hooks
 - Add post-command support for preflight checks
+- Add cargo test command
 
 ## [0.7.0] - 2026-08-03