+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Docs
+
+- Update changelog configuration to categorize CI and docs
+
 ## [0.8.0] - 2026-08-04
 
 ### Added
@@ -333,6 +339,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Update .forgejo/workflows/build.yaml
 - Update .forgejo/workflows/build.yaml
 
+[unreleased]: https://github.com/jonahmakowski/cli-tool/compare/v0.8.0..HEAD
 [0.8.0]: https://github.com/jonahmakowski/cli-tool/compare/v0.7.0..v0.8.0
 [0.7.0]: https://github.com/jonahmakowski/cli-tool/compare/v0.6.6..v0.7.0
 [0.6.6]: https://github.com/jonahmakowski/cli-tool/compare/v0.6.5-beta6..v0.6.6